+++ CHANGELOG.md
@@ -42,6 +42,7 @@ of Doomseeker.
 or "MacOS" (application bundle), the binary directory will no longer be added to
 the wad file paths by default.
 - Updated the internal zlib library to 1.2.11. This affects Windows builds.
+- Logs will now be printed to stderr instead of stdout.
 - Linux: Also try to look for WADs in /usr/[local]/share/games/doom.
 - Linux: Explicit "Copy (text)" actions shall store the data both in keyboard
 and mouse selection clipboards, allowing it to be pasted with either keyboard